Move Canonicalize into resolve. Rename File to FilePath and have more consistent naming.
